Add extractor docs
