+++ docs/extractors/__category__.yaml
@@ -1,0 +1,2 @@
+label: 'Extractors'
+position: 600
