+++ src/connectors/impls/kv.rs
@@ -14,10 +14,7 @@
 
 // #![cfg_attr(coverage, no_coverage)]
 use crate::{
-    codec::{
-        json::{Json, Sorted},
-        Codec,
-    },
+    codec::{json::Json, json_sorted::Sorted, Codec},
     connectors::prelude::*,
     errors::err_connector_def,
 };
