+++ tremor-script/src/ast/raw.rs
@@ -32,9 +32,9 @@ use crate::{
         UnaryOpKind,
     },
     errors::{err_generic, error_generic, error_missing_effector, Kind as ErrorKind, Result},
+    extractor::Extractor,
     impl_expr, impl_expr_exraw, impl_expr_no_lt,
     prelude::*,
-    tilde::Extractor,
     KnownKey, Value,
 };
 pub use base_expr::BaseExpr;
